Move canSearchItemByEffectiveLocation() from ItemEffectiveLocationTest into ItemStorageTest.
Also extract these methods into TestBaseWithInventoryUtil:
* @BeforeClass beforeAny()
* createItem(JsonObject itemToCreate)
* createItem(Item item)
* buildItem(UUID holdingsRecordId, UUID permLocation, UUID tempLocation)
