+++ src/test/java/org/folio/rest/api/ItemStorageTest.java
@@ -5,17 +5,13 @@ import static org.folio.rest.support.JsonObjectMatchers.validationErrorMatches;
 import static org.folio.rest.support.http.InterfaceUrls.holdingsStorageUrl;
 import static org.folio.rest.support.http.InterfaceUrls.instancesStorageUrl;
 import static org.folio.rest.support.http.InterfaceUrls.itemsStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.loanTypesStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.locCampusStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.locInstitutionStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.locLibraryStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.locationsStorageUrl;
-import static org.folio.rest.support.http.InterfaceUrls.materialTypesStorageUrl;
 import static org.folio.util.StringUtil.urlEncode;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.hasItem;
+import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.core.IsNull.notNullValue;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
@@ -31,17 +27,18 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import java.util.stream.Collectors;
 
+import org.folio.rest.jaxrs.model.Item;
+import org.folio.rest.jaxrs.model.Items;
 import org.folio.rest.support.AdditionalHttpStatusCodes;
+import org.folio.rest.support.IndividualResource;
 import org.folio.rest.support.JsonArrayHelper;
 import org.folio.rest.support.JsonErrorResponse;
 import org.folio.rest.support.Response;
 import org.folio.rest.support.ResponseHandler;
-import org.folio.rest.support.client.LoanTypesClient;
-import org.folio.rest.support.client.MaterialTypesClient;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 import io.vertx.core.json.JsonArray;
@@ -51,39 +48,7 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
 
   private static final String TAG_VALUE = "test-tag";
 
-  private static String journalMaterialTypeID;
-  private static String bookMaterialTypeID;
-  private static String canCirculateLoanTypeID;
-  private static UUID mainLibraryLocationId;
-  private static UUID annexLibraryLocationId;
-
-  @BeforeClass
-  public static void beforeAny()
-    throws MalformedURLException,
-    InterruptedException,
-    ExecutionException,
-    TimeoutException {
-
-    StorageTestSuite.deleteAll(itemsStorageUrl(""));
-    StorageTestSuite.deleteAll(holdingsStorageUrl(""));
-    StorageTestSuite.deleteAll(instancesStorageUrl(""));
-
-    StorageTestSuite.deleteAll(materialTypesStorageUrl(""));
-    StorageTestSuite.deleteAll(locationsStorageUrl(""));
-    StorageTestSuite.deleteAll(locLibraryStorageUrl(""));
-    StorageTestSuite.deleteAll(locCampusStorageUrl(""));
-    StorageTestSuite.deleteAll(locInstitutionStorageUrl(""));
-    StorageTestSuite.deleteAll(loanTypesStorageUrl(""));
-
-    MaterialTypesClient materialTypesClient = new MaterialTypesClient(client, materialTypesStorageUrl(""));
-    journalMaterialTypeID = materialTypesClient.create("journal");
-    bookMaterialTypeID = materialTypesClient.create("book");
-    canCirculateLoanTypeID = new LoanTypesClient(client, loanTypesStorageUrl("")).create("Can Circulate");
-
-    LocationsTest.createLocUnits(true);
-    mainLibraryLocationId = LocationsTest.createLocation(null, "Main Library (Item)", "It/M");
-    annexLibraryLocationId = LocationsTest.createLocation(null, "Annex Library (item)", "It/A");
-  }
+  // see also @BeforeClass TestBaseWithInventoryUtil.beforeAny()
 
   @Before
   public void beforeEach() {
@@ -960,6 +925,48 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void canSearchItemByEffectiveLocation() throws Exception {
+    UUID holdingsWithPermLocation = createInstanceAndHolding(mainLibraryLocationId);
+    UUID holdingsWithTempLocation = createInstanceAndHolding(mainLibraryLocationId, annexLibraryLocationId);
+
+    Item itemWithHoldingPermLocation = buildItem(holdingsWithPermLocation, null, null);
+    Item itemWithHoldingTempLocation = buildItem(holdingsWithTempLocation, null, null);
+    Item itemWithTempLocation = buildItem(holdingsWithPermLocation, onlineLocationId, null);
+    Item itemWithPermLocation = buildItem(holdingsWithTempLocation, null, secondFloorLocationId);
+    Item itemWithAllLocation = buildItem(holdingsWithTempLocation, secondFloorLocationId, onlineLocationId);
+
+    Item[] itemsToCreate = {itemWithHoldingPermLocation, itemWithHoldingTempLocation,
+      itemWithTempLocation, itemWithPermLocation, itemWithAllLocation};
+
+    for (Item item : itemsToCreate) {
+      IndividualResource createdItem = createItem(item);
+      assertTrue(createdItem.getJson().containsKey("effectiveLocationId"));
+    }
+
+    Items mainLibraryItems = findItems("effectiveLocationId=" + mainLibraryLocationId);
+    Items annexLibraryItems = findItems("effectiveLocationId=" + annexLibraryLocationId);
+    Items onlineLibraryItems = findItems("effectiveLocationId=" + onlineLocationId);
+    Items secondFloorLibraryItems = findItems("effectiveLocationId=" + secondFloorLocationId);
+
+    assertEquals(1, mainLibraryItems.getTotalRecords().intValue());
+    assertThat(mainLibraryItems.getItems().get(0).getId(), is(itemWithHoldingPermLocation.getId()));
+
+    assertEquals(1, annexLibraryItems.getTotalRecords().intValue());
+    assertThat(annexLibraryItems.getItems().get(0).getId(), is(itemWithHoldingTempLocation.getId()));
+
+    assertEquals(2, onlineLibraryItems.getTotalRecords().intValue());
+
+    assertThat(onlineLibraryItems.getItems()
+        .stream()
+        .map(Item::getId)
+        .collect(Collectors.toList()),
+      hasItems(itemWithTempLocation.getId(), itemWithAllLocation.getId()));
+
+    assertEquals(1, secondFloorLibraryItems.getTotalRecords().intValue());
+    assertThat(secondFloorLibraryItems.getItems().get(0).getId(), is(itemWithPermLocation.getId()));
+  }
+
+  @Test
   public void cannotSearchForItemsUsingADefaultField()
     throws MalformedURLException,
     InterruptedException,
@@ -1094,22 +1101,6 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
     return getCompleted.get(5, TimeUnit.SECONDS);
   }
 
-  private void createItem(JsonObject itemToCreate)
-    throws MalformedURLException,
-    InterruptedException,
-    ExecutionException,
-    TimeoutException {
-
-    CompletableFuture<Response> createCompleted = new CompletableFuture<>();
-
-    client.post(itemsStorageUrl(""), itemToCreate, StorageTestSuite.TENANT_ID,
-      ResponseHandler.json(createCompleted));
-
-    Response response = createCompleted.get(2, TimeUnit.SECONDS);
-
-    assertThat(response.getStatusCode(), is(201));
-  }
-
   private JsonObject createItemRequest(
       UUID id,
       UUID holdingsRecordId,
@@ -1166,6 +1157,16 @@ public class ItemStorageTest extends TestBaseWithInventoryUtil {
 
   private JsonObject interestingTimes(UUID itemId, UUID holdingsRecordId) {
     return createItemRequest(itemId, holdingsRecordId, "56454543534");
+  }
+
+  private Items findItems(String searchQuery) throws Exception {
+    CompletableFuture<Response> searchCompleted = new CompletableFuture<>();
+
+    client.get(itemsStorageUrl("?query=") + urlEncode(searchQuery),
+      StorageTestSuite.TENANT_ID, ResponseHandler.json(searchCompleted));
+
+    return searchCompleted.get(5, TimeUnit.SECONDS).getJson()
+      .mapTo(Items.class);
   }
 
   private JsonObject addTags(String tagValue, UUID holdingsRecordId) {
