+++ src/test/java/org/folio/rest/api/TestBaseWithInventoryUtil.java
@@ -1,11 +1,33 @@
 package org.folio.rest.api;
 
+import static org.folio.rest.support.http.InterfaceUrls.holdingsStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.instancesStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.itemsStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.loanTypesStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.locCampusStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.locInstitutionStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.locLibraryStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.locationsStorageUrl;
+import static org.folio.rest.support.http.InterfaceUrls.materialTypesStorageUrl;
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
 import java.net.MalformedURLException;
+import java.util.Random;
 import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import org.folio.rest.jaxrs.model.Item;
+import org.folio.rest.support.IndividualResource;
+import org.folio.rest.support.Response;
+import org.folio.rest.support.ResponseHandler;
 import org.folio.rest.support.builders.HoldingRequestBuilder;
+import org.folio.rest.support.client.LoanTypesClient;
+import org.folio.rest.support.client.MaterialTypesClient;
+import org.junit.BeforeClass;
 
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -15,6 +37,32 @@ import io.vertx.core.json.JsonObject;
  * @author ne
  */
 public abstract class TestBaseWithInventoryUtil extends TestBase {
+  protected static final String PERMANENT_LOCATION_ID_KEY = "permanentLocationId";
+  protected static final String TEMPORARY_LOCATION_ID_KEY = "temporaryLocationId";
+  protected static final String EFFECTIVE_LOCATION_ID_KEY = "effectiveLocationId";
+
+  protected static final String MAIN_LIBRARY_LOCATION = "Main Library";
+  protected static final String SECOND_FLOOR_LOCATION = "Second Floor";
+  protected static final String ANNEX_LIBRARY_LOCATION = "Annex Library";
+  protected static final String ONLINE_LOCATION = "Online";
+  protected static final String THIRD_FLOOR_LOCATION = "Third Floor";
+  protected static final String FOURTH_FLOOR_LOCATION = "Fourth Floor";
+
+  protected static UUID   journalMaterialTypeId;
+  protected static String journalMaterialTypeID;
+  protected static UUID   bookMaterialTypeId;
+  protected static String bookMaterialTypeID;
+  protected static UUID   canCirculateLoanTypeId;
+  protected static String canCirculateLoanTypeID;
+
+  // Creating the UUIDs here because they are used in ItemEffectiveLocationTest.parameters()
+  // that JUnit calls *before* the @BeforeClass beforeAny() method.
+  protected static UUID mainLibraryLocationId = UUID.randomUUID();
+  protected static UUID annexLibraryLocationId = UUID.randomUUID();
+  protected static UUID onlineLocationId = UUID.randomUUID();
+  protected static UUID secondFloorLocationId = UUID.randomUUID();
+  protected static UUID thirdFloorLocationId = UUID.randomUUID();
+  protected static UUID fourthFloorLocationId = UUID.randomUUID();
 
   // These UUIDs were taken from reference-data folder.
   // When the vertical gets started the data from the reference-data folder are loaded to the DB.
@@ -25,6 +73,41 @@ public abstract class TestBaseWithInventoryUtil extends TestBase {
   protected static final UUID UUID_TEXT = UUID.fromString("6312d172-f0cf-40f6-b27d-9fa8feaf332f");
   protected static final UUID UUID_INSTANCE_TYPE = UUID.fromString("535e3160-763a-42f9-b0c0-d8ed7df6e2a2");
 
+  @BeforeClass
+  public static void beforeAny()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+
+    StorageTestSuite.deleteAll(itemsStorageUrl(""));
+    StorageTestSuite.deleteAll(holdingsStorageUrl(""));
+    StorageTestSuite.deleteAll(instancesStorageUrl(""));
+
+    StorageTestSuite.deleteAll(materialTypesStorageUrl(""));
+    StorageTestSuite.deleteAll(locationsStorageUrl(""));
+    StorageTestSuite.deleteAll(locLibraryStorageUrl(""));
+    StorageTestSuite.deleteAll(locCampusStorageUrl(""));
+    StorageTestSuite.deleteAll(locInstitutionStorageUrl(""));
+    StorageTestSuite.deleteAll(loanTypesStorageUrl(""));
+
+    MaterialTypesClient materialTypesClient = new MaterialTypesClient(client, materialTypesStorageUrl(""));
+    journalMaterialTypeID = materialTypesClient.create("journal");
+    journalMaterialTypeId = UUID.fromString(journalMaterialTypeID);
+    bookMaterialTypeID = materialTypesClient.create("book");
+    bookMaterialTypeId = UUID.fromString(bookMaterialTypeID);
+    LoanTypesClient loanTypesClient = new LoanTypesClient(client, loanTypesStorageUrl(""));
+    canCirculateLoanTypeID = loanTypesClient.create("Can Circulate");
+    canCirculateLoanTypeId = UUID.fromString(canCirculateLoanTypeID);
+
+    LocationsTest.createLocUnits(true);
+    LocationsTest.createLocation(mainLibraryLocationId,  MAIN_LIBRARY_LOCATION,  "TestBaseWI/M");
+    LocationsTest.createLocation(annexLibraryLocationId, ANNEX_LIBRARY_LOCATION, "TestBaseWI/A");
+    LocationsTest.createLocation(onlineLocationId,       ONLINE_LOCATION,        "TestBaseWI/O");
+    LocationsTest.createLocation(secondFloorLocationId,  SECOND_FLOOR_LOCATION,  "TestBaseWI/SF");
+    LocationsTest.createLocation(thirdFloorLocationId,   THIRD_FLOOR_LOCATION,   "TestBaseWI/TF");
+    LocationsTest.createLocation(fourthFloorLocationId,  FOURTH_FLOOR_LOCATION,  "TestBaseWI/FF");
+  }
 
   protected static UUID createInstanceAndHolding(UUID holdingsPermanentLocationId)
     throws ExecutionException, InterruptedException, MalformedURLException, TimeoutException {
@@ -71,6 +154,46 @@ public abstract class TestBaseWithInventoryUtil extends TestBase {
     );
   }
 
+  protected static Item buildItem(UUID holdingsRecordId,
+                                   UUID permLocation,
+                                   UUID tempLocation) {
+    JsonObject itemToCreate = new JsonObject();
+
+    itemToCreate.put("id", UUID.randomUUID().toString());
+    itemToCreate.put("holdingsRecordId", holdingsRecordId.toString());
+    itemToCreate.put("barcode", Long.toString(new Random().nextLong()));
+    itemToCreate.put("status", new JsonObject().put("name", "Available"));
+    itemToCreate.put("materialTypeId", journalMaterialTypeID);
+    itemToCreate.put("permanentLoanTypeId", canCirculateLoanTypeID);
+    if (tempLocation != null) {
+      itemToCreate.put(TEMPORARY_LOCATION_ID_KEY, tempLocation.toString());
+    }
+    if (permLocation != null) {
+      itemToCreate.put(PERMANENT_LOCATION_ID_KEY, permLocation.toString());
+    }
+
+    return itemToCreate.mapTo(Item.class);
+  }
+
+  protected void createItem(JsonObject itemToCreate)
+      throws MalformedURLException,
+      InterruptedException,
+      ExecutionException,
+      TimeoutException {
+
+    CompletableFuture<Response> createCompleted = new CompletableFuture<>();
+
+    client.post(itemsStorageUrl(""), itemToCreate, StorageTestSuite.TENANT_ID,
+        ResponseHandler.json(createCompleted));
+
+    Response response = createCompleted.get(2, TimeUnit.SECONDS);
+
+    assertThat(response.getStatusCode(), is(201));
+  }
+
+  protected IndividualResource createItem(Item item) throws Exception {
+    return itemsClient.create(JsonObject.mapFrom(item));
+  }
 
   protected static JsonObject identifier(UUID identifierTypeId, String value) {
     return new JsonObject()
